move row inside node-list
